+++ src/infra/search/elasticsearch/src/es_helpers/mod.rs
@@ -15,8 +15,8 @@ mod es_phrase_search_policy_builder;
 mod es_rrf_combiner;
 mod es_search_query_builder;
 
-pub(crate) use es_entity_index_mappings::*;
-pub(crate) use es_entity_versioned_entity_index::*;
+pub use es_entity_index_mappings::*;
+pub use es_entity_versioned_entity_index::*;
 pub(crate) use es_highlight_extractor::*;
 pub(crate) use es_multi_match_policy_builder::*;
 pub(crate) use es_phrase_search_policy_builder::*;