+++ src/infra/search/elasticsearch/tests/mod.rs
@@ -7,15 +7,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod es_search_base_harness;
-mod es_test_context;
+#![feature(assert_matches)]
 
-pub use es_search_base_harness::*;
-pub use es_test_context::*;
-
-pub use crate::es_helpers::{
-    ElasticsearchIndexMappings,
-    ElasticsearchVersionedEntityIndex,
-    EntityIndexEnsureOutcome,
-    IndexVersionMetadata,
-};
+mod tests;